Time: 38 ms (48.39%), Space: 16.7 MB (51.61%) - LeetHub

+++ 2177-find-three-consecutive-integers-that-sum-to-a-given-number/2177-find-three-consecutive-integers-that-sum-to-a-given-number.py
@@ -6,8 +6,8 @@ class Solution:
             we are looking for (x) + (x+1) + (x+2) == num
             3x + 3 = num
             3x = num-3
-            x = (num-3)/3
-            if x is not a float, return [x, x+1, x+2] otherwise return []
+            x = (num-3) / 3
+            if x is not a decimal, return [x, x+1, x+2] otherwise return []
         '''
         
         x = (num-3) / 3
